Insert문 안되는것 확인

+++ src/main/java/com/example/youtubeapiprac/mapper/ResumeMapper.java
@@ -6,5 +6,5 @@ import org.apache.ibatis.annotations.Mapper;
 @Mapper
 public interface ResumeMapper {
 
-    int insertResume(ResumeRequest resumeRequest);
+    Long insertResume(ResumeRequest resumeRequest);
 }
